create used addresses component
