+++ packages/blockchain-wallet-v4-frontend/src/scenes/Settings/Addresses/Btc/ManageAddresses/UsedAddresses/index.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import UsedAddressesTemplate from './template'
+
+export default class UsedAddressesContainer extends React.PureComponent {
+  render () {
+    return <UsedAddressesTemplate />
+  }
+}
