+++ packages/blockchain-wallet-v4-frontend/src/scenes/Settings/Addresses/Btc/ManageAddresses/UsedAddresses/template.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const UsedAddressesTemplate = () => (
+  <div>
+    <h3>Used Addresses</h3>
+  </div>
+)
+
+export default UsedAddressesTemplate
